Feat: New command system with a few ported commands
